chore: remove last functions from utils.js
